requirements.txt
Flask==3.0.3
gspread==6.1.4
oauth2client==4.1.3
ffmpeg-python==0.2.0
python-dotenv==1.0.1
pymediainfo==6.1.0
Werkzeug==3.0.3